+++ app/src/main/java/cz/cvut/palecda1/dao/ArticleDaoRoom.kt
@@ -32,9 +32,10 @@ abstract class ArticleDaoRoom : ArticleDao {
     abstract override fun insertArticles(list: List<RoomArticle>)
 
     @Transaction
-    override fun clearAndInsertList(list: List<RoomArticle>): List<RoomArticle> {
+    override fun clearAndInsertList(list: List<RoomArticle>, delete: Boolean): List<RoomArticle> {
         val oldList = articleList()
-        deleteAll()
+        if (delete)
+            deleteAll()
         insertArticles(list)
         return list.minus(oldList)
     }
